backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5
waitress>=3.0,<4
Pillow>=10,<12
python-escpos>=3.1,<4
pytest>=8.0,<9
pytest-django>=4.8,<5
openpyxl>=3.1,<4
cryptography>=45.0.7,<47
python-dotenv>=1.0,<2
